@@ -1,9 +1,74 @@
-import { Injectable, effect, EffectRef, Injector, runInInjectionContext, inject } from '@angular/core';
+import { Injectable, effect, EffectRef, Injector, runInInjectionContext, inject, Signal } from '@angular/core';
 import { DatepickerValue } from '../utils/calendar.utils';
 
+// Type for Angular core module accessed via globalThis (for dynamic isSignal detection)
+interface AngularCoreModule {
+  isSignal?: (value: unknown) => boolean;
+}
+
+interface AngularGlobal {
+  ng?: {
+    core?: AngularCoreModule;
+  };
+}
+
+// Type for writable signal with set method
+interface WritableSignal<T> extends Signal<T> {
+  set(value: T): void;
+}
+
+// Safe isSignal detection - works with all Angular 17+ versions including 17.0.0
+// In Angular 17.0.0, isSignal might not be available, so we use function-based detection
+// We avoid importing isSignal directly to prevent build errors in Angular 17.0.0
+// This function provides compatibility across all Angular 17 sub-versions
+function safeIsSignal(value: unknown): value is Signal<unknown> {
+  if (value === null || value === undefined) {
+    return false;
+  }
+
+  // For Angular 17.1.0+, try to use isSignal if available via dynamic access
+  // We check at runtime to avoid build-time import resolution issues
+  try {
+    // Access Angular core module dynamically to check for isSignal
+    // This avoids the @angular/core/primitives/signals resolution issue in Angular 17.0.0
+    const ngCore = (globalThis as AngularGlobal).ng?.core;
+    if (ngCore?.isSignal && typeof ngCore.isSignal === 'function') {
+      return ngCore.isSignal(value);
+    }
+  } catch {
+    // isSignal not available - use fallback detection
+  }
+
+  // Fallback detection for Angular 17.0.0 and all versions
+  // This works by detecting signal-like function characteristics
+  // Signals in Angular are functions that can be called without arguments
+  if (typeof value === 'function') {
+    const fn = value as () => unknown;
+    // Signals are typically:
+    // 1. Functions without a prototype (or minimal prototype)
+    // 2. Callable without arguments
+    // 3. Return values when called
+    // In Angular 17.0.0, we treat callable functions as potential signals
+    // This is safe because we're in a reactive context (effect) where signals can be read
+    try {
+      // Check if it's callable - if it is, treat it as a potential signal
+      // This works for Angular 17.0.0 where isSignal might not be available
+      if (fn.length === 0 || fn.length === undefined) {
+        // No required arguments - likely a signal getter
+        return true;
+      }
+    } catch {
+      // Not callable - not a signal
+      return false;
+    }
+  }
+
+  return false;
+}
+
 export type SignalFormField = ({
-  value?: DatepickerValue | (() => DatepickerValue) | { (): DatepickerValue };
-  disabled?: boolean | (() => boolean) | { (): boolean };
+  value?: DatepickerValue | (() => DatepickerValue) | { (): DatepickerValue } | Signal<DatepickerValue>;
+  disabled?: boolean | (() => boolean) | { (): boolean } | Signal<boolean>;
   setValue?: (value: DatepickerValue) => void;
   updateValue?: (updater: () => DatepickerValue) => void;
 } & {
@@ -27,6 +92,100 @@ export class FieldSyncService {
   private _isUpdatingFromInternal: boolean = false;
   private readonly injector = inject(Injector);
 
+  /**
+   * Helper function to safely read a field value, handling both functions and signals
+   * This handles Angular 21 Signal Forms where field.value can be a signal directly
+   * Note: Angular signals (including computed) are detected as 'function' type
+   */
+  private readFieldValue(field: SignalFormField): DatepickerValue | null {
+    if (!field || typeof field !== 'object' || field.value === undefined) {
+      return null;
+    }
+
+    try {
+      const fieldValue = field.value;
+
+      // Explicitly check for Angular Signal (compatible with all Angular 17+ versions)
+      if (safeIsSignal(fieldValue)) {
+        const result = fieldValue() as DatepickerValue;
+        return result !== undefined ? result : null;
+      }
+
+      // If it's a function (but not a signal), call it
+      if (typeof fieldValue === 'function') {
+        try {
+          const result = fieldValue() as DatepickerValue;
+          return result !== undefined ? result : null;
+        } catch {
+          return null;
+        }
+      }
+
+      // If it's an object, it might be a signal that isSignal didn't catch (unlikely for Angular 21)
+      // or a custom signal-like object. Try calling it if it's not a Date.
+      if (fieldValue !== null && typeof fieldValue === 'object') {
+        if (fieldValue instanceof Date) {
+          return fieldValue as DatepickerValue;
+        }
+
+        // Try calling it as a last resort
+        try {
+          const result = (fieldValue as unknown as { (): DatepickerValue })() as DatepickerValue;
+          return result !== undefined ? result : null;
+        } catch {
+          // Not callable, treat as value
+          return fieldValue as DatepickerValue;
+        }
+      }
+
+      return fieldValue as DatepickerValue;
+    } catch {
+      return null;
+    }
+  }
+
+  /**
+   * Helper function to safely read disabled state
+   * Handles both direct values, functions, and signals (Angular 21 Signal Forms)
+   */
+  private readDisabledState(field: SignalFormField): boolean {
+    if (!field || typeof field !== 'object') {
+      return false;
+    }
+
+    try {
+      // Check disabled property first
+      if ('disabled' in field && field.disabled !== undefined) {
+        const disabledVal = field.disabled;
+
+        // Explicitly check for Angular Signal (compatible with all Angular 17+ versions)
+        if (safeIsSignal(disabledVal)) {
+          return !!disabledVal();
+        }
+
+        if (typeof disabledVal === 'function') {
+          return !!(disabledVal as () => boolean)();
+        }
+
+        // Handle signals detected as objects
+        if (typeof disabledVal === 'object' && disabledVal !== null) {
+          try {
+            return !!(disabledVal as unknown as { (): boolean })();
+          } catch {
+            return !!disabledVal;
+          }
+        }
+
+        return !!disabledVal;
+      }
+
+      return false;
+    } catch {
+      // Silently return false on error to allow graceful degradation
+      return false;
+    }
+  }
+
   setupFieldSync(
     field: SignalFormField,
     callbacks: FieldSyncCallbacks
@@ -39,57 +198,104 @@ export class FieldSyncService {
 
     try {
       const effectRef = runInInjectionContext(this.injector, () => effect(() => {
+        // Skip if we're updating from internal to prevent circular updates
         if (this._isUpdatingFromInternal) {
           return;
         }
-        
+
+        // Read the field value directly in the effect context
+        // This ensures all signal dependencies (including computed signals) are properly tracked
         let fieldValue: DatepickerValue | null = null;
-        
-        if (typeof field.value === 'function') {
-          try {
-            fieldValue = field.value();
-          } catch (error) {
-            console.warn('[NgxsmkDatepicker] Error reading field value:', error);
-            callbacks.onSyncError?.(error);
+
+        try {
+          const fieldValueRef = field.value;
+
+          if (fieldValueRef === undefined || fieldValueRef === null) {
             fieldValue = null;
+          } else if (safeIsSignal(fieldValueRef)) {
+            // For signals (including computed), read directly in effect context
+            // This ensures computed signals track their underlying dependencies
+            // Reading the signal here automatically tracks all its dependencies
+            const signalResult = fieldValueRef();
+            fieldValue = (signalResult !== undefined && signalResult !== null) ? signalResult as DatepickerValue : null;
+          } else if (typeof fieldValueRef === 'function') {
+            // For functions (including computed signals that isSignal might miss),
+            // call in effect context to track dependencies
+            // This handles computed signals that might not be detected by isSignal()
+            // or custom signal-like objects
+            try {
+              const funcResult = fieldValueRef();
+              // Check if the result is a signal (nested computed scenario)
+              if (safeIsSignal(funcResult)) {
+                fieldValue = (funcResult() !== undefined && funcResult() !== null) ? funcResult() as DatepickerValue : null;
+              } else {
+                fieldValue = (funcResult !== undefined && funcResult !== null) ? funcResult as DatepickerValue : null;
+              }
+            } catch {
+              fieldValue = null;
+            }
+          } else if (fieldValueRef instanceof Date) {
+            // Direct Date value
+            fieldValue = fieldValueRef as DatepickerValue;
+          } else if (typeof fieldValueRef === 'object') {
+            // Could be a range object or array
+            fieldValue = fieldValueRef as DatepickerValue;
+          } else {
+            // Fallback to helper
+            fieldValue = this.readFieldValue(field);
           }
-        } else if (field.value !== undefined) {
-          fieldValue = field.value;
+        } catch {
+          // Fallback to helper on any error
+          fieldValue = this.readFieldValue(field);
         }
-        
+
+        // Normalize the value using the same normalization function as the component
+        // This ensures consistent value comparison
         const normalizedValue = callbacks.normalizeValue(fieldValue);
-        
-        const valueChanged = !callbacks.isValueEqual(normalizedValue, this._lastKnownFieldValue as DatepickerValue);
-        const isInitialLoad = this._lastKnownFieldValue === undefined && fieldValue !== null && fieldValue !== undefined;
-        const isValueTransition = (this._lastKnownFieldValue === null || this._lastKnownFieldValue === undefined) && 
-                                 fieldValue !== null && fieldValue !== undefined;
-        
-        if (valueChanged || isInitialLoad || isValueTransition) {
+
+        // Determine if we should update
+        // Always update on initial load (even if null) to ensure component is initialized
+        const isInitialLoad = this._lastKnownFieldValue === undefined;
+
+        // Use isValueEqual to compare values - this handles date normalization and equality correctly
+        // This prevents overwriting values that were just set internally, even if effect runs after flag reset
+        const valuesAreEqual = this._lastKnownFieldValue !== undefined &&
+          callbacks.isValueEqual(normalizedValue, this._lastKnownFieldValue as DatepickerValue);
+
+        const valueChanged = !isInitialLoad && !valuesAreEqual;
+        const isValueTransition = (this._lastKnownFieldValue === null || this._lastKnownFieldValue === undefined) &&
+          fieldValue !== null && fieldValue !== undefined;
+
+        // Always update on initial load (even if null), value change, or value transition
+        // BUT skip if values are equal (prevents overwriting values set internally)
+        // This ensures the component value is always set, including null values
+        // but prevents circular updates when we just set the value internally
+        if ((isInitialLoad || valueChanged || isValueTransition) && !valuesAreEqual) {
           this._lastKnownFieldValue = normalizedValue;
           callbacks.onValueChanged(normalizedValue);
           callbacks.onCalendarGenerated?.();
           callbacks.onStateChanged?.();
-        } else if (this._lastKnownFieldValue !== fieldValue) {
-          this._lastKnownFieldValue = fieldValue as DatepickerValue;
+        } else if (!valuesAreEqual && this._lastKnownFieldValue !== normalizedValue) {
+          // Update last known value even if we don't trigger callbacks
+          // This handles edge cases where values are equal but references differ
+          // But only if values are not equal according to isValueEqual
+          this._lastKnownFieldValue = normalizedValue;
         }
-        
-        if (typeof field.disabled === 'function') {
-          try {
-            const newDisabled = field.disabled();
-            callbacks.onDisabledChanged(newDisabled);
-          } catch (error) {
-            console.warn('[NgxsmkDatepicker] Error reading field disabled state:', error);
-            callbacks.onSyncError?.(error);
-          }
-        } else if (field.disabled !== undefined) {
-          callbacks.onDisabledChanged(field.disabled);
-        }
+
+        // Always update disabled state
+        const disabled = this.readDisabledState(field);
+        callbacks.onDisabledChanged(disabled);
       }));
-      
+
       this._fieldEffectRef = effectRef;
+
+      // Effects run immediately when created, so the effect has already run
+      // and read the signal value in a reactive context. This ensures computed
+      // signals are properly tracked and all dependencies are registered.
+
       return effectRef;
     } catch (error) {
-      console.warn('[NgxsmkDatepicker] Error setting up field effect, falling back to manual sync:', error);
+      // Fall back to manual sync if effect setup fails
       callbacks.onSyncError?.(error);
       this.syncFieldValue(field, callbacks);
       return null;
@@ -98,67 +304,118 @@ export class FieldSyncService {
 
   syncFieldValue(field: SignalFormField, callbacks: FieldSyncCallbacks): boolean {
     if (!field || typeof field !== 'object') return false;
-    
-    let fieldValue: DatepickerValue | null | undefined = null;
-    try {
-      const rawValue = typeof field.value === 'function' ? field.value() : field.value;
-      fieldValue = rawValue ?? null;
-    } catch (error) {
-      console.warn('[NgxsmkDatepicker] Error reading field value during sync:', error);
-      callbacks.onSyncError?.(error);
-      return false;
-    }
-    
+
+    const fieldValue = this.readFieldValue(field);
     const normalizedValue = callbacks.normalizeValue(fieldValue);
-    
+
+    // Check if value has changed or if this is initial load
     const hasValueChanged = !callbacks.isValueEqual(normalizedValue, this._lastKnownFieldValue as DatepickerValue);
-    const isInitialLoad = this._lastKnownFieldValue === undefined && fieldValue !== null && fieldValue !== undefined;
-    const isValueTransition = (this._lastKnownFieldValue === null || this._lastKnownFieldValue === undefined) && 
-                             fieldValue !== null && fieldValue !== undefined;
-    
-    if (hasValueChanged || isInitialLoad || isValueTransition) {
+    const isInitialLoad = this._lastKnownFieldValue === undefined;
+    const isValueTransition = (this._lastKnownFieldValue === null || this._lastKnownFieldValue === undefined) &&
+      fieldValue !== null && fieldValue !== undefined;
+
+    // Always sync on initial load (even if null), value change, or value transition
+    // This ensures the component value is set correctly, including null values
+    if (isInitialLoad || hasValueChanged || isValueTransition) {
       this._lastKnownFieldValue = normalizedValue;
       callbacks.onValueChanged(normalizedValue);
       callbacks.onCalendarGenerated?.();
       callbacks.onStateChanged?.();
+
+      const disabled = this.readDisabledState(field);
+      callbacks.onDisabledChanged(disabled);
       return true;
     }
-    
-    if (this._lastKnownFieldValue !== fieldValue) {
-      this._lastKnownFieldValue = fieldValue as DatepickerValue;
+
+    // Update last known value even if we don't trigger callbacks
+    if (this._lastKnownFieldValue !== normalizedValue) {
+      this._lastKnownFieldValue = normalizedValue;
     }
-    
-    if (typeof field.disabled === 'function') {
-      try {
-        const disabled = field.disabled();
-        callbacks.onDisabledChanged(disabled);
-      } catch (error) {
-        console.warn('[NgxsmkDatepicker] Error reading field disabled state during sync:', error);
-        callbacks.onSyncError?.(error);
-      }
-    } else if (field.disabled !== undefined) {
-      callbacks.onDisabledChanged(field.disabled);
-    }
+
+    const disabled = this.readDisabledState(field);
+    callbacks.onDisabledChanged(disabled);
     return false;
   }
 
-  updateFieldFromInternal(value: DatepickerValue, field: SignalFormField): void {
-    if (!field || typeof field !== 'object') return;
+  updateFieldFromInternal(
+    value: DatepickerValue,
+    field: SignalFormField
+  ): void {
+    if (!field || typeof field !== 'object') {
+      return;
+    }
 
+    // Set flag to prevent effect from processing this update
+    // This prevents circular updates when we update the field from component
     this._isUpdatingFromInternal = true;
-    
+
     try {
-      if (field.setValue) {
-        field.setValue(value);
-      } else if (field.updateValue) {
-        field.updateValue(() => value);
+      // The value passed in should already be normalized by the component
+      // Store it as the last known value BEFORE updating the field
+      // This ensures the effect sees a match if it runs, preventing overwrites
+      const normalizedValue = value;
+      this._lastKnownFieldValue = normalizedValue;
+
+      // Try setValue first (preferred method for Angular 21 Signal Forms)
+      // This works with computed signal patterns where setValue updates the underlying signal
+      if (typeof field.setValue === 'function') {
+        try {
+          field.setValue(normalizedValue);
+          // Use microtask delay to ensure effect has time to see the flag
+          // This prevents race condition where effect runs after flag is reset
+          Promise.resolve().then(() => {
+            this._isUpdatingFromInternal = false;
+          });
+          return;
+        } catch {
+          // If setValue fails, try updateValue as fallback
+          // Don't return here, continue to try updateValue
+        }
       }
-    } catch (error) {
-      console.warn('[NgxsmkDatepicker] Error updating field value:', error);
-    } finally {
-      setTimeout(() => {
-        this._isUpdatingFromInternal = false;
-      }, 0);
+
+      // Try updateValue as alternative method
+      if (typeof field.updateValue === 'function') {
+        try {
+          field.updateValue(() => normalizedValue);
+          // Use microtask delay to ensure effect has time to see the flag
+          Promise.resolve().then(() => {
+            this._isUpdatingFromInternal = false;
+          });
+          return;
+        } catch {
+          // If updateValue also fails, continue to fallback
+        }
+      }
+
+      // Fallback: try to update the underlying signal directly if field.value is a signal
+      // This handles cases where field.value is a writable signal
+      try {
+        const val = field.value;
+        // Check if it's a writable signal (has .set method)
+        // Compatible with all Angular 17+ versions
+        if (safeIsSignal(val)) {
+          const writableSignal = val as WritableSignal<DatepickerValue>;
+          if (typeof writableSignal.set === 'function') {
+            writableSignal.set(normalizedValue);
+          }
+          // Use microtask delay to ensure effect has time to see the flag
+          Promise.resolve().then(() => {
+            this._isUpdatingFromInternal = false;
+          });
+          return;
+        }
+      } catch {
+        // Silently handle - field might not support direct signal updates
+        // This is expected for computed signals which are read-only
+      }
+
+      // If no update method worked, reset flag immediately
+      this._isUpdatingFromInternal = false;
+
+    } catch {
+      // Silently handle errors to prevent breaking the application
+      // The error might be due to readonly signals or other constraints
+      this._isUpdatingFromInternal = false;
     }
   }
 
